+++ circuits/web/loggers.py
@@ -43,8 +43,9 @@ class Logger(BaseComponent):
 
         self.logger = logger
 
-    @handler("response", priority=-0.1)
-    def response(self, response):
+    @handler("response_success")
+    def log_response(self, response_event, value):
+        response = response_event.args[0]
         self.log(response)
 
     def log(self, response):